# docs/swagger.yaml
basePath: /api
definitions:
...
  domain.User:
    properties:
      account_disabled:
        description: Expose so client knows
        type: boolean
      birthday:
        type: string
      created_at:
        type: string
      disable_reason:
        description: Reason for account disable
        type: string
      disabled_at:
        type: string
      disabled_by_user_id:
        type: integer
      email:
        type: string
      email_verified:
        type: boolean
      email_verified_at:
        type: string
      id:
        type: integer
      last_login_at:
        type: string
      name:
        type: string
      profile_image:
        type: string
      role:
        description: user, admin
        type: string
      updated_at:
        type: string
    type: object
...
  handler.ErrorResponse:
    description: Error response returned when a request fails
    properties:
      error:
        example: additional error details
        type: string
      message:
        example: Invalid credentials
        type: string
    type: object
...
  handler.MessageResponse:
    description: Generic message response
    properties:
      message:
        example: Operation completed successfully
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@actalog.com
    name: ActaLog Support
    url: https://github.com/johnzastrow/actalog/issues
  description: ActaLog is a mobile-first CrossFit workout tracker API. It provides
    endpoints for user authentication, workout logging, movement tracking, personal
    records, and administrative functions.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://actalog.com/terms/
  title: ActaLog API
  version: 0.17.0
paths:
...
  /admin/users:
    get:
      consumes:
      - application/json
      description: Get a paginated list of all users
      parameters:
      - description: Max results (default 50)
        in: query
        name: limit
        type: integer
      - description: Skip N results (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Users list with total count
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List users (Admin)
      tags:
      - admin
  /admin/users/{id}:
    delete:
      consumes:
      - application/json
      description: Permanently delete a user and all their data
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Invalid user ID or cannot delete
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user (Admin)
      tags:
      - admin
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User details
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user details (Admin)
      tags:
      - admin
  /admin/users/{id}/disable:
    post:
      consumes:
      - application/json
      description: Disable a user account with optional reason
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Optional disable reason
        in: body
        name: request
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Invalid user ID or cannot disable
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Disable user account (Admin)
      tags:
      - admin
  /admin/users/{id}/enable:
    post:
      consumes:
      - application/json
      description: Re-enable a previously disabled user account
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Enable user account (Admin)
      tags:
      - admin
...
  /admin/users/{id}/role:
    put:
      consumes:
      - application/json
      description: Change a user's role to 'user' or 'admin'
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: New role (user or admin)
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Invalid user ID or role
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change user role (Admin)
      tags:
      - admin
  /admin/users/{id}/toggle-email-verification:
    post:
      consumes:
      - application/json
      description: Manually set a user's email verification status
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Verified status (true/false)
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Toggle email verification (Admin)
      tags:
      - admin
  /admin/users/{id}/unlock:
    post:
      consumes:
      - application/json
      description: Unlock a user account that was locked due to failed login attempts
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Unlock user account (Admin)
      tags:
      - admin
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"